Strip unnecessary parenthesis from template strings
Decaf wraps *every* template string with parenthesis.  There's nowhere in the Shopify codebase that requires this.

This transform relies on odd behaviour to strip out parenthesis; I'm guessing that recast sees a new property on the node and this forces a (good) destructive change.  There's an unsupported `Node#parenthesizedExpression` property that *may* make this less haphazard, but I couldn't get it to work any more effectively than this solution.